general: fix import problems.
Moves ItemStatus and IssueItemStatus class to shared project and correct
import statements into projects.
