+++ projects/admin/src/app/circulation/patron/loan/loan.component.ts
@@ -18,11 +18,11 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { DateTranslatePipe } from '@rero/ng-core';
-import { User, UserService } from '@rero/shared';
+import { ItemStatus, User, UserService } from '@rero/shared';
 import moment from 'moment';
 import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
 import { ToastrService } from 'ngx-toastr';
-import { Item, ItemAction, ItemNoteType, ItemStatus } from 'projects/admin/src/app/classes/items';
+import { Item, ItemAction, ItemNoteType } from 'projects/admin/src/app/classes/items';
 import { ItemsService } from 'projects/admin/src/app/service/items.service';
 import { PatronService } from 'projects/admin/src/app/service/patron.service';
 import { forkJoin, Subscription } from 'rxjs';
