+++ projects/admin/src/app/routes/issues-route.ts
@@ -15,10 +15,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import { RecordSearchPageComponent, RouteInterface } from '@rero/ng-core';
+import { IssueItemStatus } from '@rero/shared';
 import { of } from 'rxjs';
 import { IssuesBriefViewComponent } from '../record/brief-view/issues-brief-view/issues-brief-view.component';
 import { BaseRoute } from './base-route';
-import { IssueItemStatus } from '../classes/items';
 
 export class IssuesRoute extends BaseRoute implements RouteInterface {
