+++ projects/shared/src/public-api.ts
@@ -21,6 +21,7 @@
 
 export * from './lib/api/user-api.service';
 export * from './lib/class/ContextSettings.interface';
+export * from './lib/class/item-status';
 export * from './lib/class/user';
 export * from './lib/pipe/contribution-format.pipe';
 export * from './lib/pipe/contribution-type.pipe';
